% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSI.R
\name{calRSI}
\alias{calRSI}
\title{Radiosensitivity Index}
\usage{
calRSI(x)
}
\arguments{
\item{x}{An gene expression matrix with rownames<-gene symbol and colnames<-sample ID}
}
\value{
RSI
}
\description{
Radiosensitivity Index
}
