% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{default_query_function}
\title{Deprecated functions}
\usage{
default_query_function(
  bbox,
  stac_source,
  collection,
  start_date,
  end_date,
  limit,
  ...
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions have been deprecated in favor of better approaches.
\itemize{
\item \code{default_query_function()} was renamed to \code{rsi_query_api()}. These
functions are identical, and the older name will be removed in a future
release.
}
}
\keyword{internal}
