% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_platforms}
\alias{filter_platforms}
\alias{filter_bands}
\title{Filter indices based on (relatively) complicated fields}
\usage{
filter_platforms(
  indices = spectral_indices(),
  platforms = unique(unlist(spectral_indices()$platforms)),
  operand = c("all", "any")
)

filter_bands(
  indices = spectral_indices(),
  bands = unique(unlist(spectral_indices()$bands)),
  operand = c("all", "any")
)
}
\arguments{
\item{indices}{The data frame to filter. Must contain the relevant column.}

\item{platforms, bands}{Names of the instruments (for \code{platforms}) or spectra
(for \code{bands}) indices must contain.}

\item{`operand`}{A function defining how to apply this filter.
For instance, \code{operand = all} means that the index must contain all the
\code{platforms} or \code{bands} provided, while \code{operand = any} means that the index
must contain at least one of the \code{platforms} or \code{bands} provided.}
}
\value{
A filtered version of \code{indices}.
}
\description{
Filter indices based on (relatively) complicated fields
}
\examples{
filter_platforms(platforms = "Sentinel-2")
filter_platforms(platforms = c("Landsat-OLI", "Sentinel-2"))
filter_bands(bands = c("R", "N"), operand = any)

}
