% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{get_dem}
\alias{get_dem}
\title{Download DEMs from Planetary Computer (or potentially another STAC source)}
\usage{
get_dem(
  aoi,
  ...,
  pixel_x_size = 30,
  pixel_y_size = 30,
  buffer = 90,
  output_filename = paste0(targets::tar_random_name(), ".tif"),
  start_date = NULL,
  end_date = NULL,
  stac_source = "https://planetarycomputer.microsoft.com/api/stac/v1",
  collections = c("cop-dem-glo-30", "cop-dem-glo-90", "alos-dem", "nasadem"),
  asset_names = "data",
  download_function = download_planetary_computer,
  limit = 999,
  gdalbuildvrt_options = c("-r", "bilinear", "-overwrite"),
  gdalwarp_options = c("-r", "bilinear", "-multi", "-overwrite", "-co",
    "COMPRESS=DEFLATE", "-co", "PREDICTOR=3", "-co", "NUM_THREADS=ALL_CPUS")
)
}
\arguments{
\item{aoi}{An sf(c) object outlining the area of interest to get imagery for.
Will be used to get the bounding box used for calculating metrics and the
output data's CRS.}

\item{...}{Arguments passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}.}

\item{pixel_x_size}{Numeric of length 1: size of pixels in x-direction
(longitude / easting).}

\item{pixel_y_size}{Numeric of length 1: size of pixels in y-direction
(latitude / northing).}

\item{buffer}{An optional buffer (in the same units as \code{aoi}) to apply to the
AOI before downloading data.}

\item{output_filename}{The filename to write the output raster to.}

\item{start_date, end_date}{Character of length 1: The first and last date,
respectively, of imagery to include in metrics calculations. Should be in
YYYY-MM-DD format.}

\item{stac_source}{Character of length 1: the STAC URL to download
imagery from. Defaults to the \code{stac_source} attribute of \code{remap_band_names}.}

\item{collections}{a \code{character} vector of collection IDs to include in
the search for items. Only items in one of the provided collections will be
searched.}

\item{asset_names}{The names of the assets to download.}

\item{download_function}{A function that takes the output from
\code{\link[rstac:stac_search]{rstac::stac_search()}} and executes the request. Use this if you need to sign
requests for your endpoint, or use a non-GET verb.}

\item{limit}{an \code{integer} defining the maximum number of results
to return. If not informed, it defaults to the service implementation.}

\item{gdalbuildvrt_options, gdalwarp_options}{Options passed to \code{gdalbuildvrt}
and \code{gdalwarp}, respectively, through the \code{options} argument of
\code{\link[sf:gdal_utils]{sf::gdal_utils()}}.}
}
\value{
\code{output_filename}, unchanged.
}
\description{
This function downloads DEMs directly from a STAC endpoint, by default
Microsoft's Planetary Computer. It would likely work with other endpoints,
so long as \code{stac_source}, \code{collections}, \code{asset_names}, and
\code{download_function} are provided, but currently all default and input
checking are specific to Planetary Computer.
}
\examples{
\dontrun{
aoi <- sf::st_point(c(-74.912131, 44.080410)) |>
  sf::st_sfc() |>
  sf::st_set_crs(4326) |>
  sf::st_transform(3857) |>
  sf::st_buffer(100)

get_dem(aoi)
}

}
