% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landsat_imagery.R
\name{get_landsat_imagery}
\alias{get_landsat_imagery}
\title{Get Landsat imagery}
\usage{
get_landsat_imagery(
  aoi,
  start_date,
  end_date,
  ...,
  pixel_x_size = 30,
  pixel_y_size = 30,
  imagery_time_step = "P1D",
  imagery_aggregation_function = "median",
  imagery_resampling_function = "bilinear",
  reduce_time = TRUE,
  remap_band_names = landsat_band_mapping$planetary_computer_v1,
  stac_source = attr(remap_band_names, "stac_source"),
  collections = attr(remap_band_names, "collection_name"),
  qa_band_name = attr(remap_band_names, "qa_name"),
  platforms = c("landsat-9", "landsat-8"),
  download_function = attr(remap_band_names, "download_function"),
  limit = 999,
  creation_options = list(COMPRESS = "DEFLATE", PREDICTOR = "3"),
  reduce_function = "median",
  directory = "."
)
}
\arguments{
\item{aoi}{An sf(c) object outlining the area of interest to get imagery for.
Will be used to get the bounding box used for calculating metrics and the
output data's CRS.}

\item{start_date, end_date}{Character of length 1: The first and last date,
respectively, of imagery to include in metrics calculations. Should be in
YYYY-MM-DD format.}

\item{...}{These dots are for future extensions and must be empty.}

\item{pixel_x_size, pixel_y_size}{Numeric of length 1: size of pixels in
x-direction (longitude / easting) and y-direction (latitude / northing).}

\item{stac_source}{Character of length 1: the STAC URL to download
imagery from.}

\item{platforms}{The acceptable satellites to download imagery from.}

\item{limit}{an \code{integer} defining the maximum number of results
to return. If not informed, it defaults to the service implementation.}
}
\value{
\code{output_filename}, unchanged.
}
\description{
This function downloads Landsat imagery from a STAC endpoint, by default
Microsoft's Planetary Computer. The \code{landsat_band_mapping} object included
in this package contains metadata relevant to the Planetary Computer STAC
endpoint.
}
\examples{
\dontrun{
aoi <- sf::st_point(c(-74.912131, 44.080410)) |>
  sf::st_sfc() |>
  sf::st_set_crs(4326) |>
  sf::st_transform(3857) |>
  sf::st_buffer(100)

get_landsat_imagery(
  aoi,
  "2022-06-01",
  "2022-08-30"
)
}

}
