% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentinel1_imagery.R
\name{get_sentinel1_imagery}
\alias{get_sentinel1_imagery}
\title{Retrieve composites of Sentinel-1 images from STAC endpoints}
\usage{
get_sentinel1_imagery(
  aoi,
  start_date,
  end_date,
  ...,
  pixel_x_size = 10,
  pixel_y_size = 10,
  buffer = 90,
  output_file = paste0(proceduralnames::make_english_names(1), ".tif"),
  reduce_function = "median",
  asset_names = sentinel1_band_mapping$planetary_computer_v1,
  stac_source = attr(asset_names, "stac_source"),
  collections = attr(asset_names, "collection_name"),
  download_function = attr(asset_names, "download_function"),
  limit = 999,
  gdalwarp_options = c("-r", "bilinear", "-multi", "-overwrite", "-co",
    "COMPRESS=DEFLATE", "-co", "PREDICTOR=2", "-co", "NUM_THREADS=ALL_CPUS")
)
}
\arguments{
\item{aoi}{An sf(c) object outlining the area of interest to get imagery for.
Will be used to get the bounding box used for calculating metrics and the
output data's CRS.}

\item{start_date, end_date}{Character of length 1: The first and last date,
respectively, of imagery to include in metrics calculations. Should be in
YYYY-MM-DD format.}

\item{...}{Arguments passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}.}

\item{pixel_x_size}{Numeric of length 1: size of pixels in x-direction
(longitude / easting).}

\item{pixel_y_size}{Numeric of length 1: size of pixels in y-direction
(latitude / northing).}

\item{buffer}{An optional buffer (in the same units as \code{aoi}) to apply to the
AOI before downloading data.}

\item{reduce_function}{The (quoted) name of a function from
\code{terra} (for instance, \link[terra:summarize-generics]{terra::median}) used to combine downloaded images
into a single composite.}

\item{asset_names}{The names of the assets to download.}

\item{stac_source}{Character of length 1: the STAC URL to download
imagery from. Defaults to the \code{stac_source} attribute of \code{remap_band_names}.}

\item{collections}{a \code{character} vector of collection IDs to include in
the search for items. Only items in one of the provided collections will be
searched.}

\item{download_function}{A function that takes the output from
\code{\link[rstac:stac_search]{rstac::stac_search()}} and executes the request. Use this if you need to sign
requests for your endpoint, or use a non-GET verb.}

\item{limit}{an \code{integer} defining the maximum number of results
to return. If not informed, it defaults to the service implementation.}
}
\value{
\code{output_file}, unchanged.
}
\description{
This function retrieves composites of Sentinel-1 images from STAC endpoints.
The interface is similar to, but not identical to, \code{\link[=get_dem]{get_dem()}} and
\code{\link[=get_landsat_imagery]{get_landsat_imagery()}}. Importantly, where \code{\link[=get_dem]{get_dem()}} does not average
images at all and \code{\link[=get_landsat_imagery]{get_landsat_imagery()}} provides the \code{reduce_time} boolean
to control if raw images or composites should be returned, this function
\emph{always} produces composites of all the images within the spatiotemporal area
of interest.
}
\details{
Both the GRD and RTC collections are supported. To download RTC data, set
\code{collections} to \code{sentinel-1-rtc}, and supply your subscription key as an
environment variable named \code{lignin_pc_key} (through, e.g., \code{Sys.setenv()}
or your \code{.Renviron} file).
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aoi <- sf::st_point(c(-74.912131, 44.080410)) |>
  sf::st_sfc() |>
  sf::st_set_crs(4326) |>
  sf::st_transform(5070) |>
  sf::st_buffer(100)

get_sentinel1_imagery(
  aoi,
  "2022-06-01",
  "2022-08-30"
)
\dontshow{\}) # examplesIf}
}
