% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landsat_band_mapping}
\alias{landsat_band_mapping}
\title{Landsat2 band mapping}
\format{
An object of class \code{list} of length 1.
}
\usage{
landsat_band_mapping
}
\description{
This object is a named list, with names corresponding to Landsat band
names and values corresponding to band names in \code{spectral_indices}.
}
\keyword{datasets}
