% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_functions.R
\name{landsat_mask_function}
\alias{landsat_mask_function}
\title{Create a Landsat mask raster from the QA band}
\usage{
landsat_mask_function(raster)
}
\arguments{
\item{raster}{The QA band of a Landsat image}
}
\value{
A boolean raster to be used to mask a Landsat image
}
\description{
Create a Landsat mask raster from the QA band
}
