% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_platform_filter.R
\name{landsat_platform_filter}
\alias{landsat_platform_filter}
\title{Filter Landsat features to only specific platforms}
\usage{
landsat_platform_filter(items, platforms)
}
\arguments{
\item{items}{A \code{STACItemCatalog} containing some number of features}

\item{platforms}{A vector of acceptable platforms, for instance \code{landsat-9}.
Note that this refers to satellite names, and \emph{not} to platforms in
\code{spectral_indices()}.}
}
\description{
Filter Landsat features to only specific platforms
}
