% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_planetary_computer.R
\name{query_planetary_computer}
\alias{query_planetary_computer}
\title{Execute STAC queries against the Planetary Computer}
\usage{
query_planetary_computer(q, subscription_key = Sys.getenv("rsi_pc_key"))
}
\arguments{
\item{q}{The query object from \code{\link[rstac:stac_search]{rstac::stac_search()}}.}

\item{subscription_key}{Optionally, a subscription key associated with your
Planetary Computer account. At the time of writing, this is required for
downloading Sentinel 1 RTC products, as well as NAIP imagery. This key will
be automatically used if the environment variable \code{rsi_pc_key} is set.}
}
\description{
Execute STAC queries against the Planetary Computer
}
