% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsi-package.R
\docType{package}
\name{rsi-package}
\alias{rsi}
\alias{rsi-package}
\title{rsi: Utilities for Retrieving Satellite Imagery, Calculating Spectral Indices, and Wrangling the Outputs}
\description{
Tools for downloading spatial data from spatiotemporal asset catalogs ('STAC') and computing standard spectral indices from raster data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Permian-Global-Research/rsi}
  \item Report bugs at \url{https://github.com/Permian-Global-Research/rsi/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

Other contributors:
\itemize{
  \item Permian Global [copyright holder, funder]
}

}
\keyword{internal}
