% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rsi_gdal_options}
\alias{rsi_gdal_options}
\alias{rsi_gdal_config_options}
\alias{rsi_gdalwarp_options}
\title{Default options for GDAL}
\usage{
rsi_gdal_config_options()

rsi_gdalwarp_options()
}
\value{
A vector of options for GDAL commands.
}
\description{
These functions provide useful default options for GDAL functions,
making downloading and warping (hopefully!) more efficient for
most use cases.
}
