% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sentinel2_band_mapping}
\alias{sentinel2_band_mapping}
\title{Sentinel-2 band mapping}
\format{
An object of class \code{list} of length 3.
}
\usage{
sentinel2_band_mapping
}
\description{
This object is a named list, with names corresponding to Sentinel-2 band
names and values corresponding to band names in \code{spectral_indices}.
}
\keyword{datasets}
