% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_functions.R
\name{sentinel2_mask_function}
\alias{sentinel2_mask_function}
\title{Create a Sentinel-2 mask raster from the SCL band}
\usage{
sentinel2_mask_function(raster)
}
\arguments{
\item{raster}{The SCL band of a Sentinel-2 image}
}
\value{
A boolean raster to be used to mask a Sentinel-2 image
}
\description{
Create a Sentinel-2 mask raster from the SCL band
}
