% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.R
\name{spectral_indices_url}
\alias{spectral_indices_url}
\title{Get the URL to download spectral indices from}
\usage{
spectral_indices_url()
}
\value{
A URL to download indices from.
}
\description{
Get the URL to download spectral indices from
}
\examples{
spectral_indices_url()

}
