% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_rasters.R
\name{stack_rasters}
\alias{stack_rasters}
\title{Create and save a multi-band VRT by combining input rasters}
\usage{
stack_rasters(
  rasters,
  raster_path,
  ...,
  resolution,
  extent,
  reference_raster = 1,
  resampling_method = "bilinear",
  band_names
)
}
\arguments{
\item{rasters}{A list of rasters to combine into a single multi-band raster,
either as SpatRaster objects from \code{\link[terra:rast]{terra::rast()}} or character file paths
to files that can be read by \code{\link[terra:rast]{terra::rast()}}. Rasters will be "stacked" upon
one another, preserving values. They must share CRS.}

\item{raster_path}{The location to save the final "stacked" raster. Must be
a VRT file.}

\item{...}{These dots are for future extensions and must be empty.}

\item{resolution}{Numeric of length 2, representing the target X and Y
resolution of the output raster. If only a single value is provided, it will
be used for both X and Y resolution; if more than 2 values are provided, an
error is thrown.}

\item{extent}{Numeric of length 4, representing the target xmin, ymin, xmax,
and ymax values of the output raster (its bounding box), in that order.}

\item{reference_raster}{The position (index) of the raster in \code{rasters} to
take extent, resolution, and CRS information from. No reprojection is done.
If \code{resolution} or \code{extent} are provided, they override the values from the
reference raster.}

\item{resampling_method}{The method to use when resampling to different
resolutions.}

\item{band_names}{Either a character vector of band names, or a function that
when given a character vector of band names, returns a character vector of
the same length containing new band names.}
}
\value{
\code{raster_path}, unchanged.
}
\description{
This function creates a VRT that "stacks" all the bands of its input rasters,
as though they were loaded one after another into a GIS. The VRT is fast
to create and does not require much space, but does require the input rasters
not be moved or altered. Run
\code{sf::gdal_utils("warp", raster_path, some_path.tif)} to turn the output VRT
into a standalone TIF file.
}
\examples{
stack_rasters(
  list(
    system.file("rasters/ta.tif", package = "lignin"),
    system.file("rasters/ta.tif", package = "lignin")
  ),
  tempfile(fileext = ".vrt")
)

}
