% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsi-package.R
\docType{package}
\name{rsi-package}
\alias{rsi}
\alias{rsi-package}
\title{rsi: Utilities for Retrieving Satellite Imagery, Calculating Spectral Indices, and Wrangling the Outputs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for downloading spatial data from 'spatiotemporal' asset catalogs ('STAC') and computing standard spectral indices from raster data. Methods focus on interoperability with the broader spatial ecosystem; function arguments and outputs use classes from 'sf' and 'terra', and data downloading functions support complex 'CQL2' queries using 'rstac'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Permian-Global-Research/rsi}
  \item \url{https://permian-global-research.github.io/rsi/}
  \item Report bugs at \url{https://github.com/Permian-Global-Research/rsi/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

Other contributors:
\itemize{
  \item Permian Global [copyright holder, funder]
}

}
\keyword{internal}
