% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsi-package.R
\docType{package}
\name{rsi-package}
\alias{rsi}
\alias{rsi-package}
\title{rsi: Efficiently Retrieve and Process Satellite Imagery}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Downloads spatial data from spatiotemporal asset catalogs ('STAC'), computes standard spectral indices from the Awesome Spectral Indices project (Montero et al. (2023) \doi{10.1038/s41597-023-02096-0}) against raster data, and glues the outputs together into predictor bricks. Methods focus on interoperability with the broader spatial ecosystem; function arguments and outputs use classes from 'sf' and 'terra', and data downloading functions support complex 'CQL2' queries using 'rstac'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Permian-Global-Research/rsi}
  \item \url{https://permian-global-research.github.io/rsi/}
  \item Report bugs at \url{https://github.com/Permian-Global-Research/rsi/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

Other contributors:
\itemize{
  \item Felipe Carvalho (Felipe reviewed the package (v. 0.3.0) for rOpenSci, see <https://github.com/ropensci/software-review/issues/636>) [reviewer]
  \item Michael Sumner (Michael reviewed the package (v. 0.3.0) for rOpenSci, see <https://github.com/ropensci/software-review/issues/636>) [reviewer]
  \item Permian Global [copyright holder, funder]
}

}
\keyword{internal}
